refactor: create a single analyzer for both cli and plugin
