+++ example/example.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:dart_code_metrics/metrics_analyzer.dart';
 import 'package:dart_code_metrics/reporters.dart';
+import 'package:dart_code_metrics/src/analyzers/lint_analyzer/lint_analyzer.dart';
 
 Future<void> main() async {
   // Get some folder you would like to analyze
@@ -25,22 +26,8 @@ Future<void> main() async {
     antiPatterns: {'long-method': {}},
   );
 
-  // Store keeps reported issues in format-agnostic way
-  final store = MetricsRecordsStore.store();
-
-  // Analyzer traverses files and report its findings to passed store
-  final analyzer = MetricsAnalyzer(store, config);
-
-  // Runner coordinates analyzer and store
-  final runner = MetricsAnalysisRunner(
-    analyzer,
-    store,
-    foldersToAnalyze,
-    rootFolder,
-  );
-
-  // Execute run() to analyze files and collect results
-  await runner.run();
+  final result = await const LintAnalyzer()
+      .runCliAnalysis(foldersToAnalyze, rootFolder, config);
 
   // Now runner.results() contains some insights about analyzed code. Let's report it!
   // For a simple example we would report results to terminal
@@ -49,7 +36,7 @@ Future<void> main() async {
   final reporter = LintConsoleReporter(stdout);
 
   // Now pass collected analysis reports from runner to reporter and that's it
-  await reporter.report(runner.results());
+  await reporter.report(result);
 
   // There is also JsonReporter for making machine-readable reports
   // HtmlReporter produces fancy html-documents with bells and whistles
