+++ lib/metrics_analyzer.dart
@@ -1,6 +1,3 @@
 export 'package:dart_code_metrics/src/config_builder/models/config.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/constants.dart';
-export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_analysis_runner.dart';
-export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_analyzer.dart';
-export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_records_builder.dart';
-export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_records_store.dart';
+export 'package:dart_code_metrics/src/analyzers/lint_analyzer/lint_analyzer.dart';
