+++ lib/src/analyzers/lint_analyzer/anti_patterns/anti_patterns_list/long_parameter_list.dart
@@ -4,9 +4,9 @@ import '../../../models/issue.dart';
 import '../../../models/scoped_function_declaration.dart';
 import '../../metrics/metric_utils.dart';
 import '../../metrics/metrics_list/number_of_parameters_metric.dart';
-import '../../metrics_analyzer_utils.dart';
 import '../models/obsolete_pattern.dart';
 import '../pattern_utils.dart' as utils;
+import '../pattern_utils.dart';
 
 class LongParameterList extends ObsoletePattern {
   static const String patternId = 'long-parameter-list';
