+++ lib/src/analyzers/lint_analyzer/anti_patterns/pattern_utils.dart
@@ -4,6 +4,7 @@ import 'package:source_span/source_span.dart';
 
 import '../../models/internal_resolved_unit_result.dart';
 import '../../models/issue.dart';
+import '../../models/scoped_function_declaration.dart';
 import '../../models/severity.dart';
 import 'models/obsolete_pattern.dart';
 
@@ -40,4 +41,18 @@ Issue createIssue(
     message: message,
     verboseMessage: recommendation,
   );
+}
+
+int getArgumentsCount(ScopedFunctionDeclaration dec) {
+  final declaration = dec.declaration;
+
+  int? argumentsCount;
+  if (declaration is FunctionDeclaration) {
+    argumentsCount =
+        declaration.functionExpression.parameters?.parameters.length;
+  } else if (declaration is MethodDeclaration) {
+    argumentsCount = declaration.parameters?.parameters.length;
+  }
+
+  return argumentsCount ?? 0;
 }
