+++ lib/src/analyzers/lint_analyzer/parserd_config.dart
@@ -1,25 +1,27 @@
 import 'package:glob/glob.dart';
 import 'package:meta/meta.dart';
 
-import '../analyzers/lint_analyzer/anti_patterns/models/obsolete_pattern.dart';
-import '../analyzers/lint_analyzer/metrics/models/metric.dart';
-import '../analyzers/lint_analyzer/rules/models/rule.dart';
+import 'anti_patterns/models/obsolete_pattern.dart';
+import 'metrics/models/metric.dart';
+import 'rules/models/rule.dart';
 
 @immutable
-class AnalyzerPluginConfig {
+class ParsedConfig {
   final Iterable<Glob> globalExcludes;
   final Iterable<Rule> codeRules;
+  final Iterable<ObsoletePattern> antiPatterns;
+  final Iterable<Metric> classesMetrics;
   final Iterable<Metric> methodsMetrics;
   final Iterable<Glob> metricsExcludes;
-  final Iterable<ObsoletePattern> antiPatterns;
   final Map<String, Object> metricsConfig;
 
-  const AnalyzerPluginConfig(
+  const ParsedConfig(
     this.globalExcludes,
     this.codeRules,
+    this.antiPatterns,
+    this.classesMetrics,
     this.methodsMetrics,
     this.metricsExcludes,
-    this.antiPatterns,
     this.metricsConfig,
   );
 }
