+++ lib/src/utils/exclude_utils.dart
@@ -1,9 +1,8 @@
-import 'package:analyzer/dart/analysis/results.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart' as p;
 
-bool isExcluded(AnalysisResult result, Iterable<Glob> excludes) {
-  final path = result.path?.replaceAll(r'\', '/');
+bool isExcluded(String? absolutePath, Iterable<Glob> excludes) {
+  final path = absolutePath?.replaceAll(r'\', '/');
 
   return path != null && excludes.any((exclude) => exclude.matches(path));
 }
