+++ test/analyzers/lint_analyzer/metrics_analysis_recorder_test.dart
@@ -1,322 +1,322 @@
-@TestOn('vm')
-import 'package:analyzer/dart/ast/ast.dart';
-import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/metrics_list/number_of_parameters_metric.dart';
-import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_analysis_recorder.dart';
-import 'package:dart_code_metrics/src/analyzers/models/class_type.dart';
-import 'package:dart_code_metrics/src/analyzers/models/function_type.dart';
-import 'package:dart_code_metrics/src/analyzers/models/issue.dart';
-import 'package:dart_code_metrics/src/analyzers/models/replacement.dart';
-import 'package:dart_code_metrics/src/analyzers/models/scoped_class_declaration.dart';
-import 'package:dart_code_metrics/src/analyzers/models/scoped_function_declaration.dart';
-import 'package:dart_code_metrics/src/analyzers/models/severity.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:source_span/source_span.dart';
-import 'package:test/test.dart';
+// @TestOn('vm')
+// import 'package:analyzer/dart/ast/ast.dart';
+// import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/metrics_list/number_of_parameters_metric.dart';
+// import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_analysis_recorder.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/class_type.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/function_type.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/issue.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/replacement.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/scoped_class_declaration.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/scoped_function_declaration.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/severity.dart';
+// import 'package:mocktail/mocktail.dart';
+// import 'package:source_span/source_span.dart';
+// import 'package:test/test.dart';
 
-import '../../stubs_builders.dart';
+// import '../../stubs_builders.dart';
 
-class ClassDeclarationMock extends Mock implements ClassDeclaration {}
+// class ClassDeclarationMock extends Mock implements ClassDeclaration {}
 
-class FunctionDeclarationMock extends Mock implements FunctionDeclaration {}
+// class FunctionDeclarationMock extends Mock implements FunctionDeclaration {}
 
-class SimpleIdentifierMock extends Mock implements SimpleIdentifier {}
+// class SimpleIdentifierMock extends Mock implements SimpleIdentifier {}
 
-void main() {
-  group('MetricsAnalysisRecorder', () {
-    const filePath = '/home/developer/work/project/example.dart';
-    const rootDirectory = '/home/developer/work/project/';
+// void main() {
+//   group('MetricsAnalysisRecorder', () {
+//     const filePath = '/home/developer/work/project/example.dart';
+//     const rootDirectory = '/home/developer/work/project/';
 
-    group('recordFile', () {
-      group('Stores component records for file', () {
-        const componentName = 'simpleClass';
+//     group('recordFile', () {
+//       group('Stores component records for file', () {
+//         const componentName = 'simpleClass';
 
-        final simpleIdentifierMock = SimpleIdentifierMock();
-        when(() => simpleIdentifierMock.name).thenReturn(componentName);
+//         final simpleIdentifierMock = SimpleIdentifierMock();
+//         when(() => simpleIdentifierMock.name).thenReturn(componentName);
 
-        final classDeclarationMock = ClassDeclarationMock();
-        when(() => classDeclarationMock.name).thenReturn(simpleIdentifierMock);
+//         final classDeclarationMock = ClassDeclarationMock();
+//         when(() => classDeclarationMock.name).thenReturn(simpleIdentifierMock);
 
-        final record =
-            ScopedClassDeclaration(ClassType.generic, classDeclarationMock);
+//         final record =
+//             ScopedClassDeclaration(ClassType.generic, classDeclarationMock);
 
-        test('Stores record for file', () {
-          final componentRecord = buildComponentRecordStub();
+//         test('Stores record for file', () {
+//           final componentRecord = buildComponentRecordStub();
 
-          expect(
-            MetricsAnalysisRecorder()
-                .recordFile(filePath, rootDirectory, (b) {
-                  b.recordClass(record, componentRecord);
-                })
-                .records()
-                .single
-                .classes,
-            containsPair(componentName, componentRecord),
-          );
-        });
-      });
+//           expect(
+//             MetricsAnalysisRecorder()
+//                 .recordFile(filePath, rootDirectory, (b) {
+//                   b.recordClass(record, componentRecord);
+//                 })
+//                 .records()
+//                 .single
+//                 .classes,
+//             containsPair(componentName, componentRecord),
+//           );
+//         });
+//       });
 
-      group('Stores function records for file', () {
-        const functionName = 'simpleFunction';
+//       group('Stores function records for file', () {
+//         const functionName = 'simpleFunction';
 
-        final simpleIdentifierMock = SimpleIdentifierMock();
-        when(() => simpleIdentifierMock.name).thenReturn(functionName);
+//         final simpleIdentifierMock = SimpleIdentifierMock();
+//         when(() => simpleIdentifierMock.name).thenReturn(functionName);
 
-        final functionDeclarationMock = FunctionDeclarationMock();
-        when(() => functionDeclarationMock.name)
-            .thenReturn(simpleIdentifierMock);
+//         final functionDeclarationMock = FunctionDeclarationMock();
+//         when(() => functionDeclarationMock.name)
+//             .thenReturn(simpleIdentifierMock);
 
-        final record = ScopedFunctionDeclaration(
-          FunctionType.function,
-          functionDeclarationMock,
-          null,
-        );
+//         final record = ScopedFunctionDeclaration(
+//           FunctionType.function,
+//           functionDeclarationMock,
+//           null,
+//         );
 
-        test('Stores record for file', () {
-          final functionRecord = buildFunctionRecordStub(
-            location: SourceSpanBase(
-              SourceLocation(0, line: 1),
-              SourceLocation(0, line: 2),
-              '',
-            ),
-            metrics: [
-              buildMetricValueStub<int>(
-                id: NumberOfParametersMetric.metricId,
-                value: 3,
-              ),
-            ],
-          );
+//         test('Stores record for file', () {
+//           final functionRecord = buildFunctionRecordStub(
+//             location: SourceSpanBase(
+//               SourceLocation(0, line: 1),
+//               SourceLocation(0, line: 2),
+//               '',
+//             ),
+//             metrics: [
+//               buildMetricValueStub<int>(
+//                 id: NumberOfParametersMetric.metricId,
+//                 value: 3,
+//               ),
+//             ],
+//           );
 
-          expect(
-            MetricsAnalysisRecorder()
-                .recordFile(filePath, rootDirectory, (b) {
-                  b.recordFunctionData(record, functionRecord);
-                })
-                .records()
-                .single
-                .functions,
-            containsPair(functionName, functionRecord),
-          );
-        });
-      });
-      group('Stores issues for file', () {
-        test('Aggregates issues for file', () {
-          const _issueRuleId = 'ruleId1';
-          const _issueRuleDocumentation = 'https://docu.edu/ruleId1.html';
-          const _issueMessage = 'first issue message';
-          const _issueCorrection = 'correction';
-          const _issueCorrectionComment = 'correction comment';
+//           expect(
+//             MetricsAnalysisRecorder()
+//                 .recordFile(filePath, rootDirectory, (b) {
+//                   b.recordFunctionData(record, functionRecord);
+//                 })
+//                 .records()
+//                 .single
+//                 .functions,
+//             containsPair(functionName, functionRecord),
+//           );
+//         });
+//       });
+//       group('Stores issues for file', () {
+//         test('Aggregates issues for file', () {
+//           const _issueRuleId = 'ruleId1';
+//           const _issueRuleDocumentation = 'https://docu.edu/ruleId1.html';
+//           const _issueMessage = 'first issue message';
+//           const _issueCorrection = 'correction';
+//           const _issueCorrectionComment = 'correction comment';
 
-          final issueRecord = MetricsAnalysisRecorder()
-              .recordFile(filePath, rootDirectory, (b) {
-                b.recordIssues([
-                  Issue(
-                    ruleId: _issueRuleId,
-                    documentation: Uri.parse(_issueRuleDocumentation),
-                    severity: Severity.style,
-                    location: SourceSpanBase(
-                      SourceLocation(
-                        1,
-                        sourceUrl: Uri.parse(filePath),
-                        line: 2,
-                        column: 3,
-                      ),
-                      SourceLocation(6, sourceUrl: Uri.parse(filePath)),
-                      'issue',
-                    ),
-                    message: _issueMessage,
-                    suggestion: const Replacement(
-                      comment: _issueCorrectionComment,
-                      replacement: _issueCorrection,
-                    ),
-                  ),
-                ]);
-              })
-              .records()
-              .single
-              .issues
-              .single;
+//           final issueRecord = MetricsAnalysisRecorder()
+//               .recordFile(filePath, rootDirectory, (b) {
+//                 b.recordIssues([
+//                   Issue(
+//                     ruleId: _issueRuleId,
+//                     documentation: Uri.parse(_issueRuleDocumentation),
+//                     severity: Severity.style,
+//                     location: SourceSpanBase(
+//                       SourceLocation(
+//                         1,
+//                         sourceUrl: Uri.parse(filePath),
+//                         line: 2,
+//                         column: 3,
+//                       ),
+//                       SourceLocation(6, sourceUrl: Uri.parse(filePath)),
+//                       'issue',
+//                     ),
+//                     message: _issueMessage,
+//                     suggestion: const Replacement(
+//                       comment: _issueCorrectionComment,
+//                       replacement: _issueCorrection,
+//                     ),
+//                   ),
+//                 ]);
+//               })
+//               .records()
+//               .single
+//               .issues
+//               .single;
 
-          expect(issueRecord.ruleId, _issueRuleId);
-          expect(
-            issueRecord.documentation.toString(),
-            _issueRuleDocumentation,
-          );
-          expect(issueRecord.message, _issueMessage);
-          expect(issueRecord.suggestion!.comment, _issueCorrectionComment);
-          expect(issueRecord.suggestion!.replacement, _issueCorrection);
-        });
-      });
-    });
+//           expect(issueRecord.ruleId, _issueRuleId);
+//           expect(
+//             issueRecord.documentation.toString(),
+//             _issueRuleDocumentation,
+//           );
+//           expect(issueRecord.message, _issueMessage);
+//           expect(issueRecord.suggestion!.comment, _issueCorrectionComment);
+//           expect(issueRecord.suggestion!.replacement, _issueCorrection);
+//         });
+//       });
+//     });
 
-    group('recordComponent', () {
-      const componentName = 'simpleClass';
+//     group('recordComponent', () {
+//       const componentName = 'simpleClass';
 
-      final simpleIdentifierMock = SimpleIdentifierMock();
-      when(() => simpleIdentifierMock.name).thenReturn(componentName);
+//       final simpleIdentifierMock = SimpleIdentifierMock();
+//       when(() => simpleIdentifierMock.name).thenReturn(componentName);
 
-      final classDeclarationMock = ClassDeclarationMock();
-      when(() => classDeclarationMock.name).thenReturn(simpleIdentifierMock);
+//       final classDeclarationMock = ClassDeclarationMock();
+//       when(() => classDeclarationMock.name).thenReturn(simpleIdentifierMock);
 
-      final record =
-          ScopedClassDeclaration(ClassType.generic, classDeclarationMock);
+//       final record =
+//           ScopedClassDeclaration(ClassType.generic, classDeclarationMock);
 
-      test('store record for file', () {
-        final componentRecord = buildComponentRecordStub();
+//       test('store record for file', () {
+//         final componentRecord = buildComponentRecordStub();
 
-        final recorder = MetricsAnalysisRecorder()
-            .recordFile(filePath, rootDirectory, (recorder) {
-          recorder.recordClass(record, componentRecord);
-        });
+//         final recorder = MetricsAnalysisRecorder()
+//             .recordFile(filePath, rootDirectory, (recorder) {
+//           recorder.recordClass(record, componentRecord);
+//         });
 
-        expect(
-          recorder.records().single.classes,
-          containsPair(componentName, componentRecord),
-        );
-      });
-    });
+//         expect(
+//           recorder.records().single.classes,
+//           containsPair(componentName, componentRecord),
+//         );
+//       });
+//     });
 
-    group('recordFunction', () {
-      const functionName = 'simpleFunction';
+//     group('recordFunction', () {
+//       const functionName = 'simpleFunction';
 
-      final simpleIdentifierMock = SimpleIdentifierMock();
-      when(() => simpleIdentifierMock.name).thenReturn(functionName);
+//       final simpleIdentifierMock = SimpleIdentifierMock();
+//       when(() => simpleIdentifierMock.name).thenReturn(functionName);
 
-      final functionDeclarationMock = FunctionDeclarationMock();
-      when(() => functionDeclarationMock.name).thenReturn(simpleIdentifierMock);
+//       final functionDeclarationMock = FunctionDeclarationMock();
+//       when(() => functionDeclarationMock.name).thenReturn(simpleIdentifierMock);
 
-      final record = ScopedFunctionDeclaration(
-        FunctionType.function,
-        functionDeclarationMock,
-        null,
-      );
+//       final record = ScopedFunctionDeclaration(
+//         FunctionType.function,
+//         functionDeclarationMock,
+//         null,
+//       );
 
-      test('store record for file', () {
-        final functionRecord = buildFunctionRecordStub(
-          location: SourceSpanBase(
-            SourceLocation(0, line: 1),
-            SourceLocation(0, line: 2),
-            '',
-          ),
-          metrics: [
-            buildMetricValueStub<int>(
-              id: NumberOfParametersMetric.metricId,
-              value: 3,
-            ),
-          ],
-        );
+//       test('store record for file', () {
+//         final functionRecord = buildFunctionRecordStub(
+//           location: SourceSpanBase(
+//             SourceLocation(0, line: 1),
+//             SourceLocation(0, line: 2),
+//             '',
+//           ),
+//           metrics: [
+//             buildMetricValueStub<int>(
+//               id: NumberOfParametersMetric.metricId,
+//               value: 3,
+//             ),
+//           ],
+//         );
 
-        final recorder = MetricsAnalysisRecorder()
-            .recordFile(filePath, rootDirectory, (recorder) {
-          recorder.recordFunctionData(record, functionRecord);
-        });
+//         final recorder = MetricsAnalysisRecorder()
+//             .recordFile(filePath, rootDirectory, (recorder) {
+//           recorder.recordFunctionData(record, functionRecord);
+//         });
 
-        expect(
-          recorder.records().single.functions,
-          containsPair(functionName, functionRecord),
-        );
-      });
-    });
+//         expect(
+//           recorder.records().single.functions,
+//           containsPair(functionName, functionRecord),
+//         );
+//       });
+//     });
 
-    group('recordDesignIssues', () {
-      test('throws StateError if we call them in invalid state', () {
-        expect(
-          () {
-            MetricsAnalysisRecorder().recordAntiPatternCases([]);
-          },
-          throwsStateError,
-        );
-      });
+//     group('recordDesignIssues', () {
+//       test('throws StateError if we call them in invalid state', () {
+//         expect(
+//           () {
+//             MetricsAnalysisRecorder().recordAntiPatternCases([]);
+//           },
+//           throwsStateError,
+//         );
+//       });
 
-      test('aggregate issues for file', () {
-        const _issuePatternId = 'patternId1';
-        const _issuePatternDocumentation = 'https://docu.edu/patternId1.html';
-        const _issueMessage = 'first pattern message';
-        const _issueRecommendation = 'recommendation';
+//       test('aggregate issues for file', () {
+//         const _issuePatternId = 'patternId1';
+//         const _issuePatternDocumentation = 'https://docu.edu/patternId1.html';
+//         const _issueMessage = 'first pattern message';
+//         const _issueRecommendation = 'recommendation';
 
-        final recorder = MetricsAnalysisRecorder()
-            .recordFile(filePath, rootDirectory, (recorder) {
-          recorder.recordAntiPatternCases([
-            Issue(
-              ruleId: _issuePatternId,
-              documentation: Uri.parse(_issuePatternDocumentation),
-              location: SourceSpanBase(
-                SourceLocation(
-                  1,
-                  sourceUrl: Uri.parse(filePath),
-                  line: 2,
-                  column: 3,
-                ),
-                SourceLocation(6, sourceUrl: Uri.parse(filePath)),
-                'issue',
-              ),
-              severity: Severity.none,
-              message: _issueMessage,
-              verboseMessage: _issueRecommendation,
-            ),
-          ]);
-        });
+//         final recorder = MetricsAnalysisRecorder()
+//             .recordFile(filePath, rootDirectory, (recorder) {
+//           recorder.recordAntiPatternCases([
+//             Issue(
+//               ruleId: _issuePatternId,
+//               documentation: Uri.parse(_issuePatternDocumentation),
+//               location: SourceSpanBase(
+//                 SourceLocation(
+//                   1,
+//                   sourceUrl: Uri.parse(filePath),
+//                   line: 2,
+//                   column: 3,
+//                 ),
+//                 SourceLocation(6, sourceUrl: Uri.parse(filePath)),
+//                 'issue',
+//               ),
+//               severity: Severity.none,
+//               message: _issueMessage,
+//               verboseMessage: _issueRecommendation,
+//             ),
+//           ]);
+//         });
 
-        final issue = recorder.records().single.antiPatternCases.single;
-        expect(issue.ruleId, _issuePatternId);
-        expect(
-          issue.documentation.toString(),
-          _issuePatternDocumentation,
-        );
-        expect(issue.message, _issueMessage);
-        expect(issue.verboseMessage, _issueRecommendation);
-      });
-    });
+//         final issue = recorder.records().single.antiPatternCases.single;
+//         expect(issue.ruleId, _issuePatternId);
+//         expect(
+//           issue.documentation.toString(),
+//           _issuePatternDocumentation,
+//         );
+//         expect(issue.message, _issueMessage);
+//         expect(issue.verboseMessage, _issueRecommendation);
+//       });
+//     });
 
-    group('recordIssues', () {
-      test('throws StateError if we call them in invalid state', () {
-        expect(
-          () {
-            MetricsAnalysisRecorder().recordIssues([]);
-          },
-          throwsStateError,
-        );
-      });
+//     group('recordIssues', () {
+//       test('throws StateError if we call them in invalid state', () {
+//         expect(
+//           () {
+//             MetricsAnalysisRecorder().recordIssues([]);
+//           },
+//           throwsStateError,
+//         );
+//       });
 
-      test('aggregate issues for file', () {
-        const _issueRuleId = 'ruleId1';
-        const _issueRuleDocumentation = 'https://docu.edu/ruleId1.html';
-        const _issueMessage = 'first issue message';
-        const _issueCorrection = 'correction';
-        const _issueCorrectionComment = 'correction comment';
+//       test('aggregate issues for file', () {
+//         const _issueRuleId = 'ruleId1';
+//         const _issueRuleDocumentation = 'https://docu.edu/ruleId1.html';
+//         const _issueMessage = 'first issue message';
+//         const _issueCorrection = 'correction';
+//         const _issueCorrectionComment = 'correction comment';
 
-        final recorder = MetricsAnalysisRecorder()
-            .recordFile(filePath, rootDirectory, (recorder) {
-          recorder.recordIssues([
-            Issue(
-              ruleId: _issueRuleId,
-              documentation: Uri.parse(_issueRuleDocumentation),
-              severity: Severity.style,
-              location: SourceSpanBase(
-                SourceLocation(
-                  1,
-                  sourceUrl: Uri.parse(filePath),
-                  line: 2,
-                  column: 3,
-                ),
-                SourceLocation(6, sourceUrl: Uri.parse(filePath)),
-                'issue',
-              ),
-              message: _issueMessage,
-              suggestion: const Replacement(
-                comment: _issueCorrectionComment,
-                replacement: _issueCorrection,
-              ),
-            ),
-          ]);
-        });
+//         final recorder = MetricsAnalysisRecorder()
+//             .recordFile(filePath, rootDirectory, (recorder) {
+//           recorder.recordIssues([
+//             Issue(
+//               ruleId: _issueRuleId,
+//               documentation: Uri.parse(_issueRuleDocumentation),
+//               severity: Severity.style,
+//               location: SourceSpanBase(
+//                 SourceLocation(
+//                   1,
+//                   sourceUrl: Uri.parse(filePath),
+//                   line: 2,
+//                   column: 3,
+//                 ),
+//                 SourceLocation(6, sourceUrl: Uri.parse(filePath)),
+//                 'issue',
+//               ),
+//               message: _issueMessage,
+//               suggestion: const Replacement(
+//                 comment: _issueCorrectionComment,
+//                 replacement: _issueCorrection,
+//               ),
+//             ),
+//           ]);
+//         });
 
-        final issue = recorder.records().single.issues.single;
-        expect(issue.ruleId, _issueRuleId);
-        expect(issue.documentation.toString(), _issueRuleDocumentation);
-        expect(issue.message, _issueMessage);
-        expect(issue.suggestion!.comment, _issueCorrectionComment);
-        expect(issue.suggestion!.replacement, _issueCorrection);
-      });
-    });
-  });
-}
+//         final issue = recorder.records().single.issues.single;
+//         expect(issue.ruleId, _issueRuleId);
+//         expect(issue.documentation.toString(), _issueRuleDocumentation);
+//         expect(issue.message, _issueMessage);
+//         expect(issue.suggestion!.comment, _issueCorrectionComment);
+//         expect(issue.suggestion!.replacement, _issueCorrection);
+//       });
+//     });
+//   });
+// }
