+++ test/analyzers/lint_analyzer/metrics_analyzer_utils_test.dart
@@ -1,73 +1,73 @@
-@TestOn('vm')
-// ignore_for_file: prefer-trailing-comma
-import 'package:analyzer/dart/ast/ast.dart';
-import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_analyzer_utils.dart';
-import 'package:dart_code_metrics/src/analyzers/models/function_type.dart';
-import 'package:dart_code_metrics/src/analyzers/models/scoped_function_declaration.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:test/test.dart';
+// @TestOn('vm')
+// // ignore_for_file: prefer-trailing-comma
+// import 'package:analyzer/dart/ast/ast.dart';
+// import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics_analyzer_utils.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/function_type.dart';
+// import 'package:dart_code_metrics/src/analyzers/models/scoped_function_declaration.dart';
+// import 'package:mocktail/mocktail.dart';
+// import 'package:test/test.dart';
 
-class ConstructorDeclarationMock extends Mock
-    implements ConstructorDeclaration {}
+// class ConstructorDeclarationMock extends Mock
+//     implements ConstructorDeclaration {}
 
-class FunctionDeclarationMock extends Mock implements FunctionDeclaration {}
+// class FunctionDeclarationMock extends Mock implements FunctionDeclaration {}
 
-class FunctionExpressionMock extends Mock implements FunctionExpression {}
+// class FunctionExpressionMock extends Mock implements FunctionExpression {}
 
-class FormalParameterListMock extends Mock implements FormalParameterList {}
+// class FormalParameterListMock extends Mock implements FormalParameterList {}
 
-class NodeListMock<E extends AstNode> extends Mock implements NodeList<E> {}
+// class NodeListMock<E extends AstNode> extends Mock implements NodeList<E> {}
 
-class MethodDeclarationMock extends Mock implements MethodDeclaration {}
+// class MethodDeclarationMock extends Mock implements MethodDeclaration {}
 
-void main() {
-  group('getArgumentsCount returns arguments count of', () {
-    late FormalParameterListMock formalParameterListMock;
-    late NodeListMock<FormalParameter> nodeListMock;
+// void main() {
+//   group('getArgumentsCount returns arguments count of', () {
+//     late FormalParameterListMock formalParameterListMock;
+//     late NodeListMock<FormalParameter> nodeListMock;
 
-    setUp(() {
-      formalParameterListMock = FormalParameterListMock();
-      nodeListMock = NodeListMock<FormalParameter>();
+//     setUp(() {
+//       formalParameterListMock = FormalParameterListMock();
+//       nodeListMock = NodeListMock<FormalParameter>();
 
-      when(() => formalParameterListMock.parameters).thenReturn(nodeListMock);
-    });
+//       when(() => formalParameterListMock.parameters).thenReturn(nodeListMock);
+//     });
 
-    test('static function', () {
-      final functionDeclarationMock = FunctionDeclarationMock();
-      final functionExpressionMock = FunctionExpressionMock();
+//     test('static function', () {
+//       final functionDeclarationMock = FunctionDeclarationMock();
+//       final functionExpressionMock = FunctionExpressionMock();
 
-      when(() => functionDeclarationMock.functionExpression)
-          .thenReturn(functionExpressionMock);
-      when(() => functionExpressionMock.parameters)
-          .thenReturn(formalParameterListMock);
-      when(() => nodeListMock.length).thenReturn(1);
+//       when(() => functionDeclarationMock.functionExpression)
+//           .thenReturn(functionExpressionMock);
+//       when(() => functionExpressionMock.parameters)
+//           .thenReturn(formalParameterListMock);
+//       when(() => nodeListMock.length).thenReturn(1);
 
-      final declaration = ScopedFunctionDeclaration(
-          FunctionType.function, functionDeclarationMock, null);
+//       final declaration = ScopedFunctionDeclaration(
+//           FunctionType.function, functionDeclarationMock, null);
 
-      expect(getArgumentsCount(declaration), equals(1));
-    });
+//       expect(getArgumentsCount(declaration), equals(1));
+//     });
 
-    test('class method', () {
-      final methodDeclarationMock = MethodDeclarationMock();
+//     test('class method', () {
+//       final methodDeclarationMock = MethodDeclarationMock();
 
-      final declaration = ScopedFunctionDeclaration(
-          FunctionType.function, methodDeclarationMock, null);
+//       final declaration = ScopedFunctionDeclaration(
+//           FunctionType.function, methodDeclarationMock, null);
 
-      expect(getArgumentsCount(declaration), 0);
+//       expect(getArgumentsCount(declaration), 0);
 
-      when(() => methodDeclarationMock.parameters)
-          .thenReturn(formalParameterListMock);
-      when(() => nodeListMock.length).thenReturn(2);
+//       when(() => methodDeclarationMock.parameters)
+//           .thenReturn(formalParameterListMock);
+//       when(() => nodeListMock.length).thenReturn(2);
 
-      expect(getArgumentsCount(declaration), equals(2));
-    });
+//       expect(getArgumentsCount(declaration), equals(2));
+//     });
 
-    test('class constructor', () {
-      final declaration = ScopedFunctionDeclaration(
-          FunctionType.function, ConstructorDeclarationMock(), null);
+//     test('class constructor', () {
+//       final declaration = ScopedFunctionDeclaration(
+//           FunctionType.function, ConstructorDeclarationMock(), null);
 
-      expect(getArgumentsCount(declaration), isZero);
-    });
-  });
-}
+//       expect(getArgumentsCount(declaration), isZero);
+//     });
+//   });
+// }
